Addition and refactor of CardContainer
